# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LabelHub",
    "description": "LabelHub MVP API foundation.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_tasks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskRead"
                  },
                  "type": "array",
                  "title": "Response List Tasks Tasks Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "operationId": "create_task_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/labeler/tasks": {
      "get": {
        "tags": [
          "labeler"
        ],
        "summary": "Marketplace",
        "operationId": "marketplace_labeler_tasks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskRead"
                  },
                  "type": "array",
                  "title": "Response Marketplace Labeler Tasks Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/labeler/tasks/{task_id}/claim": {
      "post": {
        "tags": [
          "labeler"
        ],
        "summary": "Claim Task Item",
        "operationId": "claim_task_item_labeler_tasks__task_id__claim_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/labeler/assignments/{assignment_id}": {
      "get": {
        "tags": [
          "labeler"
        ],
        "summary": "Get Assignment",
        "operationId": "get_assignment_labeler_assignments__assignment_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentDetailRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/labeler/assignments/{assignment_id}/draft": {
      "put": {
        "tags": [
          "labeler"
        ],
        "summary": "Save Draft",
        "operationId": "save_draft_labeler_assignments__assignment_id__draft_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftSaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/labeler/assignments/{assignment_id}/submit": {
      "post": {
        "tags": [
          "labeler"
        ],
        "summary": "Submit Assignment",
        "operationId": "submit_assignment_labeler_assignments__assignment_id__submit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/labeler/submissions": {
      "get": {
        "tags": [
          "labeler"
        ],
        "summary": "List Own Submissions",
        "operationId": "list_own_submissions_labeler_submissions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubmissionRead"
                  },
                  "type": "array",
                  "title": "Response List Own Submissions Labeler Submissions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/review/queue": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Review Queue",
        "operationId": "review_queue_review_queue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubmissionRead"
                  },
                  "type": "array",
                  "title": "Response Review Queue Review Queue Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/review/submissions/{submission_id}": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Get Submission Detail",
        "operationId": "get_submission_detail_review_submissions__submission_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Submission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewSubmissionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/review/submissions/{submission_id}/approve": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Approve Submission",
        "operationId": "approve_submission_review_submissions__submission_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Submission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/review/submissions/{submission_id}/return": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Return Submission",
        "operationId": "return_submission_review_submissions__submission_id__return_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Submission Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/review/submissions/batch": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Batch Review",
        "operationId": "batch_review_review_submissions_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubmissionRead"
                  },
                  "type": "array",
                  "title": "Response Batch Review Review Submissions Batch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssignmentDetailRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "labeler_id": {
            "type": "string",
            "title": "Labeler Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "claimed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Claimed At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "item": {
            "$ref": "#/components/schemas/TaskItemRead"
          },
          "submission": {
            "$ref": "#/components/schemas/SubmissionRead"
          },
          "task": {
            "$ref": "#/components/schemas/TaskRead"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "item_id",
          "labeler_id",
          "status",
          "claimed_at",
          "expires_at",
          "item",
          "submission",
          "task"
        ],
        "title": "AssignmentDetailRead"
      },
...
      "BatchReviewRequest": {
        "properties": {
          "submission_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Submission Ids"
          },
          "action": {
            "type": "string",
            "pattern": "^(approve|return)$",
            "title": "Action"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "submission_ids",
          "action"
        ],
        "title": "BatchReviewRequest"
      },
...
      "ClaimRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "labeler_id": {
            "type": "string",
            "title": "Labeler Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "claimed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Claimed At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "item": {
            "$ref": "#/components/schemas/TaskItemRead"
          },
          "submission": {
            "$ref": "#/components/schemas/SubmissionRead"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "item_id",
          "labeler_id",
          "status",
          "claimed_at",
          "expires_at",
          "item",
          "submission"
        ],
        "title": "ClaimRead"
      },
      "DraftSaveRequest": {
        "properties": {
          "answer_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Answer Payload"
          }
        },
        "type": "object",
        "title": "DraftSaveRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewActionRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "ReviewActionRequest"
      },
...
      "ReviewSubmissionDetail": {
        "properties": {
          "submission": {
            "$ref": "#/components/schemas/SubmissionRead"
          },
          "task": {
            "$ref": "#/components/schemas/TaskRead"
          },
          "item": {
            "$ref": "#/components/schemas/TaskItemRead"
          }
        },
        "type": "object",
        "required": [
          "submission",
          "task",
          "item"
        ],
        "title": "ReviewSubmissionDetail"
      },
...
      "SubmissionRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "assignment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignment Id"
          },
          "labeler_id": {
            "type": "string",
            "title": "Labeler Id"
          },
          "template_schema_id": {
            "type": "string",
            "title": "Template Schema Id"
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version"
          },
          "answer_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Answer Payload"
          },
          "status": {
            "$ref": "#/components/schemas/SubmissionStatus"
          },
          "attempt": {
            "type": "integer",
            "title": "Attempt"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "item_id",
          "assignment_id",
          "labeler_id",
          "template_schema_id",
          "schema_version",
          "answer_payload",
          "status",
          "attempt",
          "submitted_at",
          "created_at",
          "updated_at"
        ],
        "title": "SubmissionRead"
      },
...
      "SubmitRequest": {
        "properties": {
          "answer_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Answer Payload"
          }
        },
        "type": "object",
        "title": "SubmitRequest"
      },
      "TaskCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "distribution_strategy": {
            "type": "string",
            "title": "Distribution Strategy",
            "default": "manual"
          },
          "quota_per_labeler": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quota Per Labeler"
          },
          "deadline_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TaskCreate"
      },
      "TaskItemRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "external_id",
          "payload",
          "status",
          "created_at"
        ],
        "title": "TaskItemRead"
      },
      "TaskRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "distribution_strategy": {
            "type": "string",
            "title": "Distribution Strategy"
          },
          "quota_per_labeler": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quota Per Labeler"
          },
          "deadline_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "status",
          "distribution_strategy",
          "quota_per_labeler",
          "deadline_at",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "TaskRead"
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "paused",
          "ended"
        ],
        "title": "TaskStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}